+++ tests/tests/wgpu-gpu/main.rs
@@ -42,6 +42,7 @@ mod occlusion_query;
 mod oob_indexing;
 mod oom;
 mod pass_ops;
+mod per_vertex;
 mod pipeline;
 mod pipeline_cache;
 mod planar_texture;
@@ -106,6 +107,7 @@ fn all_tests() -> Vec<wgpu_test::GpuTestInitializer> {
     oob_indexing::all_tests(&mut tests);
     oom::all_tests(&mut tests);
     pass_ops::all_tests(&mut tests);
+    per_vertex::all_tests(&mut tests);
     pipeline_cache::all_tests(&mut tests);
     pipeline::all_tests(&mut tests);
     planar_texture::all_tests(&mut tests);